+++ cli/types.ts
@@ -56,6 +56,7 @@ export type CanisterConfig = {
   args?: string[];
   candid?: string;
   initArg?: string;
+  wasmMemoryLimit?: number;
   "check-stable"?: {
     path: string;
     /** @deprecated Create the file with an empty actor instead. */